+++ whynot/simulators/zika/environments.py
@@ -51,7 +51,7 @@ def observation_space():
     state_dim = State.num_variables()
     state_space_low = np.zeros(state_dim)
     state_space_high = np.inf * np.ones(state_dim)
-    return spaces.Box(state_space_low, state_space_high)
+    return spaces.Box(state_space_low, state_space_high, dtype=np.float64)
 
 
 def action_space():
@@ -64,7 +64,7 @@ def action_space():
         - Indoor residual spray use.
 
     """
-    return spaces.Box(np.zeros(4), np.ones(4))
+    return spaces.Box(np.zeros(4), np.ones(4), dtype=np.float64)
 
 
 ZikaEnv = ODEEnvBuilder(
